requirements.txt
Flask==2.2.5
llama-index==0.8.3
s3fs==2023.6.0
pydantic==1.*
python-dotenv==1.0.0
